fix: various fixes
