+++ src/constants/fdp-storage-methods.ts
@@ -1,0 +1,1 @@
+export const IS_DAPP_POD_CREATED = 'isDappPodCreated'
