+++ src/services/storage/storage.service.ts
@@ -321,34 +321,6 @@ export class Storage {
     })
   }
 
-  /**
-   * Migrates existing data to match the current version
-   * @returns promise
-   */
-  public async migrate(newVersionString: string): Promise<void> {
-    const currentVersionString = await this.getStorageVersion()
-    let currentVersion: Version
-
-    try {
-      currentVersion = versionFromString(currentVersionString)
-    } catch (error) {
-      currentVersion = {
-        major: 0,
-        minor: 0,
-        patch: 0,
-      }
-    }
-
-    const accountDapps = await getObject<AccountDapps>(Storage.dappsKey, accountDappsFactory)
-    const updatedAccountDapps = migrateDapps(accountDapps, currentVersion)
-
-    if (updatedAccountDapps) {
-      await updateObject<AccountDapps>(Storage.dappsKey, updatedAccountDapps)
-    }
-
-    await this.setStorageVesion(newVersionString)
-  }
-
   private onChangeListener(changes: { [key: string]: chrome.storage.StorageChange }) {
     for (const [key, { oldValue, newValue }] of Object.entries(changes)) {
       ;(this.listeners[key] || []).forEach((listener) => listener(newValue || oldValue))
